Proteyecto terminado

+++ models/fin_reservacion.php
@@ -30,13 +30,19 @@ if(!$enlace){
         
        
        if($datos_elim){
-           
-            $elim_cliente = "DELETE FROM clientes WHERE id_habitacion = $id_habitacion";
-            $datos_elim_cliente = mysqli_query($enlace, $elim_cliente);
 
-            if($datos_elim_cliente){
-                echo "Todo eliminado";
+            $elim_rsa = "DELETE FROM claves_rsa WHERE id_cliente = (SELECT id_cliente FROM clientes WHERE cliente = '$cliente' AND id_habitacion = $id_habitacion)";
+            $datos_rsa_elim = mysqli_query($enlace, $elim_rsa);
+            if($datos_rsa_elim){
+                $elim_cliente = "DELETE FROM clientes WHERE id_habitacion = $id_habitacion";
+                $datos_elim_cliente = mysqli_query($enlace, $elim_cliente);
+    
+                if($datos_elim_cliente){
+                    echo "Todo eliminado";
+                }
             }
+
+           
        }
         
     }
